Separating Company DTOs to its own file
