+++ src/app/hooks/utils/CompanyDTO.ts
@@ -1,0 +1,8 @@
+type Company = {
+  id: string;
+  name: string;
+  date: string;
+  score: number;
+  reasoning: string;
+  color: string;
+};
